Default tools for projects and assistants - Tools selected by default but can still be changed during chat - Refactored tool selector UI so that the list of tools can be displayed without the dialog - Updated database schema
